requirements.txt
alabaster==0.7.12
amqp==2.5.2
appdirs==1.4.3
appnope==0.1.0
argon2-cffi==19.2.0
arrow==0.15.5
asgiref==3.2.2
astroid==2.3.3
attrs==19.3.0
Babel==2.8.0
backcall==0.1.0
beautifulsoup4==4.9.1
billiard==3.6.3.0
binaryornot==0.4.4
blockchain==1.4.4
celery==4.4.2
certifi==2019.9.11
cffi==1.14.0
cfgv==3.1.0
chardet==3.0.4
click==7.1.1
cookiecutter==1.7.0
coreapi==2.3.3
coreschema==0.0.4
coverage==5.0.4
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.0
Django==3.0.5
django-allauth==0.41.0
django-celery-beat==2.0.0
django-celery-results==1.2.1
django-cors-headers==3.2.1
django-coverage-plugin==1.8.0
django-crispy-forms==1.9.0
django-debug-toolbar==2.2
django-environ==0.4.5
django-extensions==2.2.9
django-filters==0.2.1
django-model-utils==4.0.0
django-redis==4.11.0
django-rest-swagger==2.2.0
django-stubs==1.5.0
django-timezone-field==4.0
django-utils==0.0.2
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
docutils==0.16
drf-custom-related-field==0.2
drf-yasg==1.17.1
ecdsa==0.15
entrypoints==0.3
enum-compat==0.0.3
environ==1.0
environs==7.4.0
factory-boy==2.12.0
Faker==4.0.2
filelock==3.0.12
flake8==3.7.9
flake8-isort==2.9.1
Flask==1.1.2
flower==0.9.4
forex-python==1.5
future==0.18.2
html5lib==1.0.1
humanize==0.5.1
identify==1.4.14
idna==2.8
imagesize==1.2.0
importlib-metadata==0.23
inflection==0.3.1
ipdb==0.13.2
ipython==7.13.0
ipython-genutils==0.2.0
isort==4.3.21
itsdangerous==1.1.0
itypes==1.1.0
jedi==0.16.0
Jinja2==2.10.3
jinja2-time==0.2.0
kiwisolver==1.1.0
kombu==4.6.8
lazy-object-proxy==1.4.3
lxml==4.5.0
Markdown==3.2.2
MarkupSafe==1.1.1
marshmallow==3.2.1
matplotlib==3.1.1
mccabe==0.6.1
more-itertools==7.2.0
mpmath==1.1.0
multitasking==0.0.9
mypy==0.770
mypy-extensions==0.4.3
nodeenv==1.3.5
numpy==1.18.3
oauthlib==3.1.0
openapi-codec==1.3.2
packaging==19.2
pandas==1.0.3
parso==0.6.2
pathspec==0.8.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.1
pluggy==0.13.1
poyo==0.5.0
pre-commit==2.2.0
prompt-toolkit==3.0.5
psycopg2-binary==2.8.4
ptyprocess==0.6.0
py==1.8.1
pyasn1==0.4.8
pycodestyle==2.5.0
pycparser==2.20
pyflakes==2.1.1
Pygments==2.6.1
PyJWT==1.7.1
pylint==2.4.4
pylint-celery==0.3
pylint-django==2.0.14
pylint-plugin-utils==0.6
pyparsing==2.4.2
pytest==5.3.5
pytest-django==3.9.0
pytest-sugar==0.9.2
python-crontab==2.4.1
python-dateutil==2.8.1
python-dotenv==0.10.3
python-jose==3.1.0
python-magic==0.4.18
python-slugify==4.0.0
python3-openid==3.1.0
pytz==2019.3
PyYAML==5.1.2
redis==3.4.1
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
ruamel.yaml==0.16.5
ruamel.yaml.clib==0.2.0
scipy==1.4.1
simplejson==3.16.0
six==1.12.0
snowballstemmer==2.0.0
soupsieve==2.0.1
Sphinx==3.0.1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sqlparse==0.3.0
sympy==1.5.1
termcolor==1.1.0
testfixtures==6.14.0
text-unidecode==1.3
toml==0.10.0
tornado==6.0.4
traitlets==4.3.3
typed-ast==1.4.1
typing-extensions==3.7.4.2
uritemplate==3.0.0
urllib3==1.25.6
vine==1.3.0
virtualenv==20.0.17
wcwidth==0.1.9
webencodings==0.5.1
Werkzeug==1.0.1
whichcraft==0.6.1
wrapt==1.11.2
yfinance==0.1.54
zipp==0.6.0